requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
asyncio>=3.4.3
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
typing-extensions>=4.0.0
colorama>=0.4.4
tqdm>=4.65.0